clickbot_archive/clickbot_v3/requirements.txt
mss>=9.0.1
numpy>=1.26.0
pyautogui>=0.9.53
Pillow>=10.1.0
scipy>=1.11.0 